Use float64 instead of float32

+++ models/sensors.go
@@ -16,9 +16,9 @@ type Sensor struct {
 // GyroSensor produces x-y-z axes angle velocity values
 type GyroSensor struct {
 	Sensor
-	AngleVelocityX float32 `json:"x_axis_angle_velocity"`
-	AngleVelocityY float32 `json:"y_axis_angle_velocity"`
-	AngleVelocityZ float32 `json:"z_axis_angle_velocity"`
+	AngleVelocityX float64 `json:"x_axis_angle_velocity"`
+	AngleVelocityY float64 `json:"y_axis_angle_velocity"`
+	AngleVelocityZ float64 `json:"z_axis_angle_velocity"`
 }
 
 func (s GyroSensor) String() string {
@@ -42,9 +42,9 @@ func (s GyroSensor) String() string {
 // AccelSensor produces x-y-z axes gravity acceleration values
 type AccelSensor struct {
 	Sensor
-	GravityAccX float32 `json:"x_axis_gravity_acceleration"`
-	GravityAccY float32 `json:"y_axis_gravity_acceleration"`
-	GravityAccZ float32 `json:"z_axis_grativy_acceleration"`
+	GravityAccX float64 `json:"x_axis_gravity_acceleration"`
+	GravityAccY float64 `json:"y_axis_gravity_acceleration"`
+	GravityAccZ float64 `json:"z_axis_grativy_acceleration"`
 }
 
 func (s AccelSensor) String() string {
@@ -68,8 +68,8 @@ func (s AccelSensor) String() string {
 // TempSensor produces temperature and humidity values
 type TempSensor struct {
 	Sensor
-	Temperature float32 `json:"temperature"`
-	Humidity    float32 `json:"humidity"`
+	Temperature float64 `json:"temperature"`
+	Humidity    float64 `json:"humidity"`
 }
 
 func (s TempSensor) String() string {
